feat: convert to a library and add ffi bridge
